Add a module to store information associated with a user after OAuth
Now, when we get the results of authentication, we save that in a map
local to the server. That means it doesn't persist across server
restarts, but it should be enough for the demo.

Next step is to look up the user info when we get a request from the
webhook and do something with it.